src/main/java/com/example/AlleDrogo/ProductRepository.java
package com.example.AlleDrogo;

import com.example.AlleDrogo.model.Product;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Repository;

import java.util.ArrayList;
import java.util.Collections;
import java.util.List;

@Repository
public class ProductRepository {

    private static final Logger log = LoggerFactory.getLogger(ProductRepository.class);

    private final List<Product> products = new ArrayList<>();

    void addProduct(Product product){
        products.add(product);
    }

    public List<Product> getAllProducts() {
        return Collections.unmodifiableList(products);
    }

}
